Abilities block encapsulation Fixes #484
