+++ characters/admin.py
@@ -18,7 +18,7 @@ from characters.models.core import (
     MeritFlawRating,
     Specialty,
 )
-from characters.models.core.ability import Ability
+from characters.models.core.ability_block import Ability
 from characters.models.core.attribute_block import Attribute
 from characters.models.core.background_block import (
     Background,
