--- characters/models/core/ability.py
@@ -1,5 +1,0 @@
-from characters.models.core.statistic import Statistic
-
-
-class Ability(Statistic):
-    type = "ability"
