+++ characters/models/mage/focus.py
@@ -1,4 +1,4 @@
-from characters.models.core.ability import Ability
+from characters.models.core.ability_block import Ability
 from characters.models.mage.resonance import Resonance
 from core.models import Model
 from django.db import models
