+++ characters/models/mage/sorcerer.py
@@ -1,6 +1,6 @@
 from collections import defaultdict
 
-from characters.models.core.ability import Ability
+from characters.models.core.ability_block import Ability
 from characters.models.core.attribute_block import Attribute
 from characters.models.mage.faction import MageFaction
 from characters.models.mage.fellowship import SorcererFellowship
