+++ characters/views/core/human.py
@@ -1,5 +1,5 @@
 from characters.models.core import Human
-from characters.models.core.ability import Ability
+from characters.models.core.ability_block import Ability
 from characters.models.core.attribute_block import Attribute
 from characters.models.core.background_block import Background
 from characters.models.core.merit_flaw_block import MeritFlaw
