+++ characters/views/mage/sorcerer.py
@@ -7,7 +7,7 @@ from characters.forms.mage.numina import (
     NuminaRitualForm,
     PsychicPathRatingFormSet,
 )
-from characters.models.core.ability import Ability
+from characters.models.core.ability_block import Ability
 from characters.models.core.archetype import Archetype
 from characters.models.core.attribute_block import Attribute
 from characters.models.core.background_block import Background, BackgroundRating
